requirements.txt
aiohttp==3.8.5
aioredis==1.3.1
aiosignal==1.3.1
alembic==1.11.1
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.2.1
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
autoflake==2.2.0
backcall==0.2.0
bcrypt==4.0.1
billiard==4.1.0
black==23.7.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.4
debugpy==1.6.7
decorator==5.1.1
dnspython==2.4.1
ecdsa==0.18.0
email-validator==2.0.0.post2
executing==1.2.0
fastapi==0.100.1
fastapi-cache2==0.2.1
fastapi-redis-cache==0.2.5
fastapi-versioning==0.10.0
flake8==6.1.0
flower==2.0.0
frozenlist==1.4.0
grafana==0.0.1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
hiredis==2.2.3
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
humanize==4.7.0
idna==3.4
iniconfig==2.0.0
install==1.3.5
ipykernel==6.25.0
ipython==8.14.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kombu==5.3.1
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
nodeenv==1.8.0
orjson==3.9.2
packaging==23.1
parso==0.8.3
passlib==1.7.4
pathspec==0.11.2
pendulum==2.1.2
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.2.0
prometheus-client==0.17.1
prometheus-fastapi-instrumentator==6.1.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pycodestyle==2.11.0
pydantic==2.1.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.4.0
pyflakes==3.1.0
Pygments==2.15.1
pyright==1.1.320
pytest==7.4.0
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
pytzdata==2020.1
pywin32==306; platform_system=="Windows"
PyYAML==6.0.1
pyzmq==25.1.0
redis==4.6.0
requests==2.31.0
rsa==4.9
sentry-sdk==1.29.2
six==1.16.0
sniffio==1.3.0
sqladmin==0.14.0
SQLAlchemy==2.0.19
sqlalchemy-admin==0.1.2
stack-data==0.6.2
starlette==0.27.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
unicorn==2.0.1.post1
urllib3==2.0.4
uvicorn==0.23.2
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
WTForms==3.0.1
yarl==1.9.2
